requirements.txt
arxiv
bs4
chromadb
ddgs
duckdb
duckduckgo-search
faiss-cpu
huggingface-hub
ipykernel
langchain
langchain-chroma
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-openai
langchain-text-splitters
langchainhub
langsmith
mysql-connector-python
numexpr
openai
pandas
pymupdf
pypdf
python-dotenv
pytube
sentence-transformers
sqlalchemy
streamlit
unstructured
validators
wikipedia
youtube-transcript-api
